fix: moved the subscription invoice flow
