+++ apps/web/.env.example
@@ -13,7 +13,7 @@ NEXT_PUBLIC_DASHBOARD_URL=
 NEXT_PUBLIC_PORTAL_URL=
 NEXT_PUBLIC_API_URL=
 NEXT_PUBLIC_DOCS_URL=
-NEXT_PUBLIC_INVOICE_URL=
+NEXT_PUBLIC_INVOICE_URL=http://invoice.localhost:3000
 MASTER_ENCRYPTION_KEY=
 ENCRYPTION_SALT=
 SUBSCRIPTION_CONTRACT_TESTNET_ID=
