+++ apps/web/app/api/subscriptions/route.ts
@@ -61,7 +61,7 @@ export const GET = apiHandler({
         failedPaymentCount: countMap[s.id] ?? 0,
         invoice_url:
           s.status === "overdue" && s.invoiceToken
-            ? `${process.env.NEXT_PUBLIC_APP_URL}/invoice/subscription/${s.invoiceToken}`
+            ? `${process.env.NEXT_PUBLIC_INVOICE_URL}/subscription/${s.invoiceToken}`
             : null,
         createdAt: s.createdAt ?? null,
         updatedAt: s.updatedAt,
